+++ component.js
@@ -1,6 +1,7 @@
 export { SYMBOLS, TRIGGERS, EVENTS } from './consts.js';
 export { AegisComponent } from './base.js';
 export { AegisInput } from './input.js';
+export { AegisView } from './view.js';
 export {
 	getBool, setBool, getInt, setInt, getFloat, setFloat, getString, setString,
 	getURL, setURL,