+++ test/components/checkbox-test.js
@@ -1,7 +1,8 @@
 import { AegisCheckbox } from '@aegisjsproject/component/checkbox.js';
-import { SYMBOLS } from '@aegisjsproject/component/consts.js';
+import { SYMBOLS, TRIGGERS } from '@aegisjsproject/component/consts.js';
 import { html } from '@aegisjsproject/core/parsers/html.js';
 import { css } from '@aegisjsproject/core/parsers/css.js';
+import { getState, setState } from '@aegisjsproject/state/state.js';
 
 const template = html`<slot name="tick"><span part="tick" class="tick"></span></slot>`;
 const styles = css`
@@ -33,6 +34,7 @@ const styles = css`
 function toggleChecked({ type, key, currentTarget }) {
 	if (! (currentTarget.disabled || currentTarget.readOnly || (type === 'keydown' && key !== ' '))) {
 		currentTarget.checked = ! currentTarget.checked;
+		setState('aegis:checkbox', currentTarget.checked);
 	}
 }
 
@@ -42,10 +44,20 @@ export class CheckboxTest extends AegisCheckbox {
 		this.addEventListener('click', toggleChecked);
 		this.addEventListener('keydown', toggleChecked);
 		this.tabIndex = '0';
+		this.checked = getState('aegis:checkbox', false);
 	}
 
 	async [SYMBOLS.render](type, data) {
-		console.log({ el: this, type, ...data });
+		switch (type) {
+			case TRIGGERS.stateChanged:
+				this.checked = data['aegis:checkbox'];
+				break;
+		}
+		// console.log({ el: this, type, ...data });
+	}
+
+	static get [SYMBOLS.observeStates]() {
+		return ['aegis:checkbox'];
 	}
 }